monalbumphoto-to-pdf/requirements.txt
asttokens==2.0.8
backcall==0.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
decorator==5.1.1
executing==1.0.0
idna==3.3
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
requests==2.28.1
six==1.16.0
stack-data==0.5.0
traitlets==5.3.0
urllib3==1.26.12
wcwidth==0.2.5
